sheet 02

+++ AUTN_Sheet02/G_Factorial.cpp
@@ -4,7 +4,7 @@ int main() {
     int t,n;
     cin>>t;
     for(int i=1;i<=t;i++){
-        int r = 1;
+        long long r = 1;
         cin>>n;
         for(int j=1;j<=n;j++){
